% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/numprocs.r
\name{num.procs}
\alias{num.procs}
\title{Number of Processors}
\usage{
num.procs()
}
\description{
Returns the total number of physical+logical processors, as
detected by omp_get_max_threads().  Will  probably agree with
\code{parallel::detectCores()}.
}

