% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{ompexamples}
\alias{c_hello}
\alias{c_primesbelow}
\alias{c_sum}
\alias{c_sweep}
\alias{f77_hello}
\alias{f77_primesbelow}
\alias{f77_sum}
\alias{f77_sweep}
\alias{f90_hello}
\alias{f90_primesbelow}
\alias{f90_sum}
\alias{f90_sweep}
\alias{ompexamples}
\alias{rcpp_hello}
\alias{rcpp_primesbelow}
\alias{rcpp_sum}
\alias{rcpp_sweep}
\title{OpenMP Examples}
\usage{
c_hello()

c_sum(x)

c_sweep(x, vec)

c_primesbelow(n)

f77_hello()

f77_sum(x)

f77_sweep(x, vec)

f77_primesbelow(n)

f90_hello()

f90_sum(x)

f90_sweep(x, vec)

f90_primesbelow(n)

rcpp_hello()

rcpp_sum(x)

rcpp_sweep(x, vec)

rcpp_primesbelow(n)
}
\arguments{
\item{x}{A numeric vector for the sum example, and a numeric matrix for
the sweep example.}

\item{vec}{A numeric vector the same length as the number of rows as x.}

\item{n}{The number from the "primes below" example, where the function
returns the total number of primes below \code{n}.}
}
\description{
There are 3 examples, each using OpenMP, across 4 implementations:
C, C++, F77, and F2003.
}
\details{
The \code{_hello()} functions are simple hello worlds.  Note that
the order of printing by the threads is not guaranteed.

The \code{_sum()} functions sum up a numeric vector.

The \code{_sweep()} functions sweep a numeric vector from a
numeric matrix, provided the vector is exactly the length of
the number of columns of the matrix.  This is equivalent to a
special case of \code{sweep(x, STATS=vec, MARGIN=1, FUN="-")}.

The \code{_primesbelow()} functions compute the number of prime
integers below a given (positive) integer.
}

