% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Romp-package.R
\docType{package}
\name{Romp-package}
\alias{Romp-package}
\title{Romp}
\description{
TODO
}
\details{
\tabular{ll}{
   Package: \tab Romp \cr
   Type: \tab Package \cr
   License: \tab 2-Clause BSD \cr
   Needs Compilation: \tab yes \cr
}
}
\author{
Drew Schmidt
}
\keyword{Package}

